==> code/05_variant_position_molecule.R <==
library(BuenColors)
source("00_functions.R")

# Main Function that processes summary counts
process_one_sample <- function(sample_id){
  
  # import redeem df
  redeemV_data <- fread(paste0(base_dir,sample_id,".Consensus.final/RawGenotypes.Sensitive.StrandBalance"))
  
  # import and select top 60 variants per variant caller
  sens_call_df <- fread(paste0("data/redeem-sensitive-calls/",sample_id,".Consensus.final_S.tsv"))
  top_hetero_variants_redeem <- sens_call_df %>% filter(totalVAF < 0.8) %>%
    arrange(desc(CellN)) %>% head(60) %>% pull(Variants)
  
  ## mgatk calls 
  mgatk_df_in <- fread(paste0("../data/mgatk-varstats/D_",sample_id,"_hg38_v20-mtMask.variant_stats.tsv.gz"))
  mgatk_df <- mgatk_df_in %>% filter(strand_correlation > 0.65 & n_cells_conf_detected >= 3 & (log10(vmr) > -2) & mean < 0.8)
  top_hetero_variants_mgatk <- mgatk_df %>% arrange(desc(n_cells_conf_detected)) %>% 
    mutate(redeem_formatted = paste0(position, "_", gsub(">", "_", nucleotide))) %>%
    pull(redeem_formatted) %>% head(60)
  
  # Now filter the redeemV data frame and annotate molecule position
  sum_stats_redeem <- redeemV_data %>% filter(V4 %in% top_hetero_variants_redeem) %>% make_position_df()
  sum_stats_mgatk <- redeemV_data %>% filter(V4 %in% top_hetero_variants_mgatk) %>% make_position_df()
  
  # do KS tests on all variants
  redeem_df_ks <- sum_stats_redeem %>% make_ks_test_df
  mgatk_df_ks <- sum_stats_mgatk %>% make_ks_test_df
  
  # write out KS statistic for all variants
  all_ks_redeem <- redeemV_data %>% filter(V4 %in% unique(sens_call_df$Variants)) %>% make_position_df() %>% make_ks_test_df()
  write.table(all_ks_redeem,
              file = paste0("../output/", sample_id, "_redeem_KSstatistic.tsv"), 
              sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
  
  mgatk_df <- mgatk_df %>% mutate(redeem_formatted = paste0(position, "_", gsub(">", "_", nucleotide))) 
  all_ks_mgatk <- redeemV_data %>% filter(V4 %in% unique(mgatk_df$redeem_formatted)) %>% make_position_df()%>% make_ks_test_df()
  write.table(all_ks_mgatk,
              file = paste0("../output/", sample_id, "_mgatk_KSstatistic.tsv"), 
              sep = "\t", quote = FALSE, row.names = FALSE, col.names = TRUE)
  
  # Create a set of variants with bias per caller
  biased_variants_redeem <- redeem_df_ks %>% filter(statistic > 0.35) %>% pull(variant)
  biased_variants_mgatk <- mgatk_df_ks %>% filter(statistic > 0.35) %>% pull(variant)
  
  # Function to plot positions of alt alleles on molecule
  make_plot <- function(var_df, var_levels, color, title) {
    n_vars = length(unique(var_df$variant))
    pP <- ggplot(var_df %>%  mutate(variant2 = factor(as.character(variant), levels = var_levels)), 
                 aes( x = rel_position)) +
      geom_histogram(color = color, fill = color, bins = 30) + 
      facet_wrap(~variant2, scales = "free_y") +
      ggtitle(paste0(sample_id, "- ",title," top variants")) + theme_bw() +
      labs(x = "Position on molecule", y = "abundance of alt allele") 
    
    ggsave(pP, file = paste0("../plots_split/", sample_id, "_",title,".png"),
           width = (n_vars)^(1/3)*4, height = (n_vars)^(1/3)*2, dpi = 200)
    title
  }
  
  pRlow <- make_plot(
    sum_stats_redeem %>% filter( !(variant %in% biased_variants_redeem)),
    top_hetero_variants_redeem, "dodgerblue3", "Redeem_KSretained")
  
  pRhigh <- make_plot(
    sum_stats_redeem %>% filter(variant %in% biased_variants_redeem),
    top_hetero_variants_redeem, "firebrick", "Redeem_KSremoved")
  
  pMlow <- make_plot(
    sum_stats_mgatk %>% filter( !(variant %in% biased_variants_mgatk)),
    top_hetero_variants_mgatk, "dodgerblue", "mgatk_KSretained")
  
  pMhigh <- make_plot(
    sum_stats_mgatk %>% filter( variant %in% biased_variants_mgatk),
    top_hetero_variants_mgatk, "red", "mgatk_KSremoved")
  
}

# Do one per donor
process_one_sample("Young1.T1.BMMC")
process_one_sample("Youn2.BMMC")
process_one_sample("Old1.BMMC")
process_one_sample("Old2.BMMC")

# See if donor-specific biases are accounted for in weighted jaccard

# Global: import mutation weights for jaccard statistic
data(CellPCT)
V.weight <- data.frame(weight=1-CellPCT$muRate)
V.weight$Variants <- paste0("Variants",gsub("_","",CellPCT$Variant))
V.weight %>% 
  filter(Variants %in% c("Variants2AG", "Variants3TA", "Variants4CT"))

# Import Y2
sample_id <- "Old2.BMMC"
redeemV_data_d <- fread(paste0(base_dir,sample_id,".Consensus.final/RawGenotypes.Sensitive.StrandBalance")) 
redeem_df <- fread(paste0("../data/redeem-sensitive-calls/",sample_id,".Consensus.final_S.tsv"))
redeemV_data_d <- redeemV_data_d %>% filter(V4 %in% redeem_df$Variants)

# count me
count_me <- redeemV_data_d %>% group_by(V2, V4) %>% summarize(count = n())
variant_meta <- count_me %>% group_by(V4) %>%
  summarize(pct_1 = mean(count == 1), max = max(count), n1 = sum(count == 1))
variant_meta %>% 
  filter(n1 > 20) %>% 
  filter(max > 30) %>%
  arrange(desc(pct_1))

# Function to make pairs of plots
process_variant <- function(var1){
  table_var <- redeemV_data_d %>%
    filter(V4 %in% var1) %>% pull(V2) %>% table() 
  
  cells1 <- names(table_var[table_var == 1])
  cells3p <- names(table_var[table_var >= 3])
  
  # Real mutation call in one cell with 21 reads and another with 3 look good
  p1 <- redeemV_data_d %>%
    filter(V4 %in% var1) %>% filter(V2  %in% cells3p) %>%
    make_position_df() %>%
    ggplot(aes( x = rel_position)) +
    geom_histogram(bins = 30) + pretty_plot(fontsize= 6) + 
    theme(plot.title = element_text(size = 6)) + labs(x = "Position on molecule", y = "count") + 
    ggtitle(paste0(sample_id, " | ", length(cells3p), " cell(s) where ",var1," has 3+ alt_alleles"))
  
  # Now look at the other ones with only 1 read
  p2 <- redeemV_data_d %>%
    filter(V4 %in% var1) %>% filter((V2 %in%cells1)) %>% 
    make_position_df() %>%
    ggplot(aes( x = rel_position)) +
    geom_histogram(bins = 30) + pretty_plot(fontsize= 6) + 
    theme(plot.title = element_text(size = 6)) + labs(x = "Position on molecule", y = "count") + 
    ggtitle(paste0(sample_id, " | ",length(cells1), " cells where ",var1," has exactly 1 alt_allele"))
  cowplot::ggsave2(cowplot::plot_grid(p1, p2, nrow = 1), 
                   width = 6, height = 2.5, file = paste0("../plots_split/individual_var/", sample_id, "_", var1, ".pdf"))
  var1
}

process_variant("15043_G_A")
process_variant("7063_T_C")
process_variant("940_A_G")
